October Architect Project by Kwadwo Danquah
This version of the game moves slightly away from the MVC model by creating Player objects to replace the Model module. This version also moves to a server so many games can be played simultaneously. To play 2 games on the same machine, two Client files have been created each connecting to one of the two ports the server is listening on.